% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_AECM.R
\name{aecm_UI}
\alias{aecm_UI}
\alias{aecm_mod}
\title{AECM Submodule}
\usage{
aecm_UI(id, title = "| AE & Conmeds |")

aecm_mod(input, output, session, uid, ADSL, ADAE, ADCM)
}
\arguments{
\item{id}{A \code{character}. The module id. Must be consistent with the inputId
in \code{callModule}.}

\item{uid}{A \code{reactive} \code{character}. A unique subject identifier.}

\item{ADSL}{A \code{reactive} \code{data.frame} of ADaM formatted Subject Level data.
This cannot be \code{NULL}.}

\item{ADAE}{A \code{reactive} \code{data.frame} of ADaM formatted Adverse Event data.}

\item{ADCM}{A \code{reactive} \code{data.frame} of ADaM formatted Concomitant Medicine data.}

\item{innput, output, session}{Shiny server arguments.}
}
\description{
Functions to enable the AE + Conmed submodule of the patient profile
}
