% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_EXLB.R
\name{exlb_UI}
\alias{exlb_UI}
\alias{exlb_mod}
\title{EXLB Submodule}
\usage{
exlb_UI(id)

exlb_mod(input, output, session, uid, ADSL, ADLB, ADEX, ADPC)
}
\arguments{
\item{id}{A \code{character}. The namespace for the module.}

\item{uid}{A \code{reactive} \code{character}. A unique subject identifier.}

\item{ADSL}{A \code{reactive} \code{data.frame} of ADaM formatted Subject Level data.
This cannot be \code{NULL}.}

\item{ADLB}{A \code{reactive} \code{data.frame} of ADaM formatted Laboratory analysis data.}

\item{ADEX}{A \code{reactive} \code{data.frame} of ADaM formatted EXposure data.}

\item{ADPC}{A \code{reactive} \code{data.frame} of ADaM formatted Pharmacokinetics data.}

\item{innput, output, session}{Shiny server arguments.}
}
\description{
Functions to enable the exposure (EX) and lab data (LB) submodule of the patient profile
}
