% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_aecm_data}
\alias{get_aecm_data}
\title{Get AECM data}
\usage{
get_aecm_data(uid, ADSL, ADAE, ADCM)
}
\arguments{
\item{uid}{A \code{character}. A subject ID available in all three datasets}

\item{ADSL}{A \code{data.frame}. The Subject Level data.}

\item{ADAE}{A \code{data.frame}. The Adverse Event data.}

\item{ADCM}{A \code{data.frame}. The Concomitent Medication data.}
}
\description{
Create a dataset from ADSL, ADAE and ADCM for a unique subject
}
\details{
The AE+CM module displays concomitant medications and adverse events for
the patient over the study period. Events can be colored by severity, grade
or action taken.
}
\seealso{
plot_aecm
}
