% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_lab_values}
\alias{get_lab_values}
\title{Get lab values}
\usage{
get_lab_values(uid, ADLB)
}
\arguments{
\item{uid}{A \code{character}. A unique subject ID available in \code{ADLB}.}

\item{ADLB}{A \code{data.frame}. The lab values}
}
\description{
Get lab values for lab overview plot
}
