% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_uid_info}
\alias{get_uid_info}
\title{Get subject info}
\usage{
get_uid_info(uid, ADSL)
}
\value{
A \code{list} with subject level information for \code{uid}.
}
\description{
Get subject level info for a participant
}
