% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_overview.R
\name{overview_modUI}
\alias{overview_modUI}
\alias{overview_mod}
\title{Overview Submodule}
\usage{
overview_modUI(id)

overview_mod(input, output, session, uid, ADSL = NULL, ADLB = NULL,
  ADAE = NULL, ADCM = NULL, ADEG = NULL, ADPC = NULL,
  ADEX = NULL)
}
\arguments{
\item{id}{A \code{character}. The namespace for the module.}

\item{uid}{A \code{reactive} \code{character}. A unique subject identifier.}

\item{ADSL}{A \code{reactive} \code{data.frame} of ADaM formatted Subject Level data.
This cannot be \code{NULL}.}

\item{ADLB}{A \code{reactive} \code{data.frame} of ADaM formatted Laboratory analysis data.}

\item{ADAE}{A \code{reactive} \code{data.frame} of ADaM formatted Adverse Event data.}

\item{ADCM}{A \code{reactive} \code{data.frame} of ADaM formatted Concomitant Medicine data.}

\item{ADEG}{A \code{reactive} \code{data.frame} of ADaM formatted Electrocardiogram data.}

\item{ADPC}{A \code{reactive} \code{data.frame} of ADaM formatted Pharmacokinetics data.}

\item{ADEX}{A \code{reactive} \code{data.frame} of ADaM formatted EXposure data.}

\item{innput, output, session}{Shiny server arguments.}
}
\description{
Functions to enable the overview submodule of the patient profile
}
\details{
Only available tables will be displayed. By default, the first ten columns
are shown. More customization will be added later.
}
