% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_profile_module.R,
%   R/patient_profile_module_UI.R
\name{patientProfile_mod}
\alias{patientProfile_mod}
\alias{patientProfile_modUI}
\title{Patient profile module}
\usage{
patientProfile_mod(input, output, session, uid = NULL, ADSL = NULL,
  ADAE = NULL, ADCM = NULL, ADLB = NULL, ADEG = NULL, ADPC = NULL,
  ADEX = NULL)

patientProfile_modUI(id)
}
\arguments{
\item{uid}{A \code{reactive} \code{character}. A unique subject identifier.}

\item{ADSL}{A \code{reactive} \code{data.frame} of ADaM formatted Subject Level data.
This cannot be \code{NULL}.}

\item{ADAE}{A \code{reactive} \code{data.frame} of ADaM formatted Adverse Event data.}

\item{ADCM}{A \code{reactive} \code{data.frame} of ADaM formatted Concomitant Medicine data.}

\item{ADLB}{A \code{reactive} \code{data.frame} of ADaM formatted Laboratory analysis data.}

\item{ADEG}{A \code{reactive} \code{data.frame} of ADaM formatted Electrocardiogram data.}

\item{ADPC}{A \code{reactive} \code{data.frame} of ADaM formatted Pharmacokinetics data.}

\item{ADEX}{A \code{reactive} \code{data.frame} of ADaM formatted EXposure data.}

\item{id}{A \code{character}. The namespace for the module.}

\item{innput, output, session}{Shiny server arguments.}
}
\description{
Server and UI functions of the patient profile module
}
\details{
Aside from \code{ADSL}, all datasets are optional. Visualization for
missing datasets will be missing. See submodules documentation for
data requirements of each tab of the participant profile module.
}
\seealso{
\code{aecm_mod}, \code{exlb_mod}, \code{labo_mod}, \code{labp_mod},
 \code{execg_mod}, \code{ecg_mod}, \code{pk_mod}
}
