% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprofile.R
\name{pprofile-pkg}
\alias{pprofile-pkg}
\title{Patient profile package}
\description{
Patient profile module and submodules for various ADAM datase for plug & play
in Novartis Shiny apps.
}
\author{
Renan Sauteraud
}
