% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showExamples.R
\name{showExamples}
\alias{showExamples}
\title{Start patient-profile app}
\usage{
showExamples(example = "patProfile", port = NULL,
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"), display.mode = c("auto",
  "normal", "showcase"))
}
\description{
Start a simple app pre-loaded with all ADaM datasets.
}
